Removed a comment and fixed picker switchcase
