+++ PRG_MAUI_Car_Register/Model/Car.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PRG_MAUI_Car_Register.Model
+{
+    class Car : Vehicle
+    {
+        public Car() : base(Type.Bil)
+        {
+
+        }
+    }
+}
