+++ PRG_MAUI_Car_Register/Model/Truck.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PRG_MAUI_Car_Register.Model
+{
+    class Truck : Vehicle
+    {
+        public Truck() : base(Type.Lastbil)
+        {
+
+        }
+    }
+}
